ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.3
google-genai==0.3.0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.19
